requirements.txt
flask
sqlalchemy
passlib
ghapi
mysql-connector-python